dynamic-todo-app/app/pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["hassanrzprop <hassanrzprop@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
dev="todos.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
